refactor(ui): move localEditorFilePath to the locality leaf module
session-diff-panel imported the helper from chat-sidebar-file-view, closing
a dependency cycle:

  chat-sidebar-content-types -> chat-sidebar-file-view -> session-diff-panel

The helper is locality policy, not file-view rendering, so it belongs beside
isLoopbackGatewayUrl in ui/src/lib/gateway-locality.ts. Its content parameter
drops the FileSidebarContent type dependency for a plain { path, root } shape,
which makes the module a true leaf both callers can share.
